refactor: 구조 변화를 위해 RacingCarGame에 모든 함수 모아두기
