--- src/components/ResultContainer.js
@@ -1,3 +1,0 @@
-class ResultContainer {}
-
-export default ResultContainer;
